+++ robot-core/src/main/java/org/obolibrary/robot/QueryOperation.java
@@ -250,9 +250,38 @@ public class QueryOperation {
    */
   public static OWLOntology convertModel(Model model, IOHelper ioHelper, String catalogPath)
       throws IOException {
-    ByteArrayOutputStream os = new ByteArrayOutputStream();
-    RDFDataMgr.write(os, model, Lang.TTL);
-    return ioHelper.loadOntology(new ByteArrayInputStream(os.toByteArray()), catalogPath);
+    return convertModel(model, ioHelper, catalogPath, false);
+  }
+
+  /**
+   * Given a Model, an IOHelper, and a path to an XML catalog, convert the model to an OWLOntology
+   * object.
+   *
+   * @param model Model to convert to OWLOntology
+   * @param ioHelper IOHelper to load ontology
+   * @param catalogPath String path to XML catalog
+   * @param useTemporaryFile whether to use a temporary file to store intermediate results or to
+   *     keep them in memory.
+   * @return OWLOntology object version of model
+   * @throws IOException on issue loading ontology
+   */
+  public static OWLOntology convertModel(
+      Model model, IOHelper ioHelper, String catalogPath, boolean useTemporaryFile)
+      throws IOException {
+    if (useTemporaryFile) {
+      final File tempFile = File.createTempFile("robot", ".owl");
+      tempFile.deleteOnExit();
+      try (final BufferedOutputStream os =
+          new BufferedOutputStream(new FileOutputStream(tempFile))) {
+        RDFDataMgr.write(os, model, Lang.TTL);
+      }
+      return ioHelper.loadOntology(
+          new BufferedInputStream(new FileInputStream(tempFile)), catalogPath);
+    } else {
+      ByteArrayOutputStream os = new ByteArrayOutputStream();
+      RDFDataMgr.write(os, model, Lang.TTL);
+      return ioHelper.loadOntology(new ByteArrayInputStream(os.toByteArray()), catalogPath);
+    }
   }
 
   /**